+++ tests/back/Pim/WorkOrganization/Specification/Workflow/Component/Normalizer/InternalApi/PublishedProductNormalizerSpec.php
@@ -117,7 +117,6 @@ final class PublishedProductNormalizerSpec extends ObjectBehavior
         $publishedProduct->setFamily(null);
         $publishedProduct->addValue(ScalarValue::value('identifier', 'an_identifier'));
         $publishedProduct->setIdentifier('an_identifier');
-        $publishedProduct->setAssociations(new ArrayCollection());
 
         $missingAssociationAdder->addMissingAssociations($publishedProduct)->shouldBeCalled();